Carved out StarMap as a separate instance.
This is further splitting of the StarMapWidget, to make really clear
what is the interconnection and requirements. It's not the point to
clean it up, only to split it.
